# Single-threaded performance improvements in forward DWT for 5-3 and 9-7 (and other improvements)