Summary of improvements:
- Switches forward 9-7 DWT to using 32-bit floating point operations instead of fixed point arithmetic 
- For 5-3 and 9-7 foward DWT, perform the vertical pass on 8 columns at a time for better cache efficiency and to be able to use SSE2 (for 5-3) and SSE (for 9-7)
- forward DWT: more efficient 5-3 and 9-7 horizontal pass processing (removal of useless bound checking)
- inverse 9-7 DWT: add multi-threading and operate on 8 columns at a time for vertical pass
- bench_dwt utility extended to support irreversible wavelet
- performance improvements in T1 encoder (~ 9%)

Combined effect of those improvements on the bench used in PR 1248 (`time bin/opj_compress -threads XXXX -n 5 -c [256,256],[256,256],[256,256],[256,256],[256,256] -t 1024,1024 -PLT  -i T36JTT_20160914T074612_B02.tif -o out.jp2 >/dev/null`)

|             | PR 1248 | Now  |
|-------------|--------|--------|
| Threads = 0 | 17.9 s | 15.3 s |
| Threads = 4 | 5.7 s  | 5.0 s  |
| Threads = 8 | 4.4 s  | 4.0 s  |

(For convenience this PR is based on top of PR 1248, so the first 2 commits "Add multithreading ..." belong to this later one)

--------------

Funded by ESA for S2-MPC project