@rouault 2020-05-24
CC @savmickael This is ready for your review