@rouault 2020-10-05
@savmickael I'll merge this soon (before end of week), unless you want to have a look before